Removed ChainConfig from trait Chain
